GREAT RESULTS!
the method Ck that creates all the combination of tuples of the frequent
items has been created.
it seems to work but i better prefer some further test.

(i need to remove the unfrequent items from currentItems after the 1st
iteration!!!!!!!!!!)

it has been implemented also the clone for the frequentItemsTable